[UPDATE]优化

+++ leyuna-disk-client/src/main/java/com/leyuna/disk/co/FileInfoCO.java
@@ -28,11 +28,11 @@ public class FileInfoCO implements Serializable {
 
     private Integer deleted;
 
-    private Long fileSize;
+    private Double fileSize;
 
     private String userId;
 
-    private Long fileSizeTotal;
+    private Double fileSizeTotal;
 
     /**
      * 文件类型：1图片、2音乐、3视频、4文档、5其他文件
